[FE] rename: 일일 플래너 컴포넌트 폴더 이동
