+++ FE/src/pages/Planner/DayPage.tsx
@@ -2,8 +2,8 @@ import React, { useEffect, useRef, useState } from "react";
 import { useLocation } from "react-router-dom";
 import Header from "@components/planner/day/Header";
 import styles from "@styles/planner/day.module.scss";
-import TimeTable from "@components/planner/day/TimeTable";
-import TodoList from "@components/planner/day/TodoList";
+import TimeTable from "@components/planner/day/todo/TimeTable";
+import TodoList from "@components/planner/day/todo/TodoList";
 import Ment from "@components/planner/day/Ment";
 import dayjs from "dayjs";
 import CustomCursor from "@components/planner/day/CustomCursor";
